+++ MInDes/MInDes/src/modules/preprocess_modules/MicrostructureInit/VoronoiStructure.h
@@ -1,8 +1,8 @@
 #pragma once
-#include "GeometryStructure.h"
 #include <string>
 #include <iostream>
 #include <limits>
+#include "GeometryStructure.h"
 namespace pf {
 	namespace voronoi_structure {
 		// -
@@ -20,25 +20,25 @@ namespace pf {
 		inline std::vector<int> voronoi_box_position = { 0, 0, 0 };
 		inline std::vector<int> voronoi_box_size = { 0, 0, 0 };
 		inline std::vector<size_t> voronoi_phi_index_range = { 0, 0 };
-		inline std::vector<float> voronoi_con;
-		inline float voronoi_temperature = 0.0;
+		inline std::vector<double> voronoi_con;
+		inline double voronoi_temperature = 0.0;
 		inline VoronoiType voronoi_type = VoronoiType::VDT_CONST;
 		inline std::vector<geometry_structure::Point> voronoi_points;
-		inline float voronoi_const_pointsDistance = -1.0;
+		inline double voronoi_const_pointsDistance = -1.0;
 		inline geometry_structure::Point voronoi_reference_dot;
-		inline float voronoi_reference_dot_distance = -1.0;
-		inline float voronoi_reference_dot_min_pointsDistance = -1.0;
-		inline float voronoi_reference_dot_max_pointsDistance = -1.0;
+		inline double voronoi_reference_dot_distance = -1.0;
+		inline double voronoi_reference_dot_min_pointsDistance = -1.0;
+		inline double voronoi_reference_dot_max_pointsDistance = -1.0;
 		inline geometry_structure::surf_func_3D voronoi_reference_surface;
-		inline float voronoi_reference_surface_distance = -1.0;
-		inline float voronoi_reference_surface_min_pointsDistance = -1.0;
-		inline float voronoi_reference_surface_max_pointsDistance = -1.0;
+		inline double voronoi_reference_surface_distance = -1.0;
+		inline double voronoi_reference_surface_min_pointsDistance = -1.0;
+		inline double voronoi_reference_surface_max_pointsDistance = -1.0;
 		inline std::vector<geometry_structure::Point> voronoi_matrix_dots;
-		inline std::vector<float> voronoi_matrix_dots_pointsDistance;
+		inline std::vector<double> voronoi_matrix_dots_pointsDistance;
 		// -
 		void generate_voronoi_structure();
 		// void generate_voronoi_structure_new_method();
 		// -
-		void generate_voronoi_structure_in_phis(std::vector<std::vector<std::vector<float>>>& aim_phi);
+		void generate_voronoi_structure_in_phis(std::vector<std::vector<std::vector<double>>>& aim_phi);
 	}
 }
